requirements.txt
Django==3.0
django-extensions==3.1.5
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
factory-boy==3.2.1
flake8==4.0.1
pylint==2.13.5
pylint-django==2.5.3
pytest==7.1.1
pytest-cov==2.9.0
pytest-django==4.5.2
python-dotenv==0.20.0